+++ server/routes/hack.js
@@ -1,20 +1,20 @@
-const express = require("express");
+const express = require('express');
 const {
   getOnlineUsers,
   saveAndUpdateUser,
   generateNotification,
-} = require("../logic/_helpers");
+} = require('../logic/_helpers');
 
 const router = express.Router();
 
-const { pettyCrime, pettyHackRouteCriterias } = require("../logic/pettyHack");
-const { crimeRouteCriterias, fightCrime } = require("../logic/crime");
+const { pettyCrime, pettyHackRouteCriterias } = require('../logic/pettyHack');
+const { crimeRouteCriterias, fightCrime } = require('../logic/crime');
 const {
   fraudHacker,
   fraudRouteCriteria,
   attackRouteCriterias,
   fightHacker,
-} = require("../logic/hack");
+} = require('../logic/hack');
 
 const getHackFeedback = (finalResult, opponent) => {
   let message;
@@ -40,20 +40,20 @@ const getHackFeedback = (finalResult, opponent) => {
   return { message, notification };
 };
 
-const User = require("../models/User");
-const Crime = require("../models/Crime");
+const User = require('../models/User');
+const Crime = require('../models/Crime');
 
 // @POST
 // PRIVATE
 // User starts interval that calls this route every x sec and commit petty crime
 
-router.post("/pettyCrime", async (req, res) => {
+router.post('/pettyCrime', async (req, res) => {
   const userId = req.user._id;
   let user;
   try {
-    user = await User.findById(userId).populate("playerStats.city", "name");
+    user = await User.findById(userId).populate('playerStats.city', 'name');
   } catch (e) {
-    console.error("error: ", e);
+    console.error('error: ', e);
     res.status(400).json({
       success: false,
       message: JSON.stringify(e),
@@ -74,7 +74,7 @@ router.post("/pettyCrime", async (req, res) => {
 
   return res.status(200).json({
     success: true,
-    message: "pettyCrime commited",
+    message: 'pettyCrime commited',
     results: results.pettyResult,
     user: results.updatedUser,
   });
@@ -84,7 +84,7 @@ router.post("/pettyCrime", async (req, res) => {
 // PRIVATE
 // Retrieves all crimes that are available
 
-router.get("/crimes", async (req, res) => {
+router.get('/crimes', async (req, res) => {
   try {
     const now = Date.now();
     const crimes = await Crime.find({ gracePeriod: { $lte: now } }).sort({
@@ -93,7 +93,7 @@ router.get("/crimes", async (req, res) => {
     });
     return res.status(200).json({
       success: true,
-      message: "Crimes loaded..",
+      message: 'Crimes loaded..',
       crimes,
     });
   } catch (err) {
@@ -104,11 +104,21 @@ router.get("/crimes", async (req, res) => {
   }
 });
 
+const generateUserFeedback = (result) => {
+  let message;
+  if (result.won) {
+    message = `SUCCESS: You gained ${result.playerGains.bitCoins} bitcoins and ${result.playerGains.exp} xp!`;
+  } else {
+    message = 'You failed..';
+  }
+  return message;
+};
+
 // @POST
 // PRIVATE
 // Commit crime route.
 
-router.post("/crimes", async (req, res) => {
+router.post('/crimes', async (req, res) => {
   const userId = req.user._id;
   const { crimeId } = req.body;
   const batteryCost = 3;
@@ -131,9 +141,11 @@ router.post("/crimes", async (req, res) => {
     .sort({ crimeType: 1, difficulty: 1 })
     .lean();
 
+  const userFeedback = generateUserFeedback(finalResult);
+
   return res.status(200).json({
     success: true,
-    message: "Crime commited..",
+    message: userFeedback,
     finalResult,
     user: finalResult.user,
     crimes,
@@ -143,7 +155,7 @@ router.post("/crimes", async (req, res) => {
 // @POST
 // PRIVATE
 // User can steal from other players.
-router.post("/fraud/:opponentId", async (req, res) => {
+router.post('/fraud/:opponentId', async (req, res) => {
   const userId = req.user._id;
   const { opponentId } = req.params;
   const batteryCost = 4;
@@ -176,8 +188,8 @@ router.post("/fraud/:opponentId", async (req, res) => {
   await generateNotification(
     user._id,
     `${message} in ${updatedUser.playerStats.city.name}`,
-    "Logs",
-    true
+    'Logs',
+    true,
   );
   const success = !!finalResult.playerGains.bitCoinStolen;
   return res.status(200).json({
@@ -192,7 +204,7 @@ router.post("/fraud/:opponentId", async (req, res) => {
 // PRIVATE
 // User can hack another plater.
 // /opponentId/attack
-router.post("/:opponentId", async (req, res) => {
+router.post('/:opponentId', async (req, res) => {
   const userId = req.user._id;
   const { opponentId } = req.params;
   const batteryCost = 6;
@@ -207,7 +219,7 @@ router.post("/:opponentId", async (req, res) => {
     user,
     opponent,
     batteryCost,
-    now
+    now,
   );
 
   if (disallowed) {
@@ -222,7 +234,7 @@ router.post("/:opponentId", async (req, res) => {
     opponent,
     batteryCost,
     now,
-    userIsOnline
+    userIsOnline,
   );
 
   const updatedUser = await saveAndUpdateUser(finalResult.user);
@@ -234,8 +246,8 @@ router.post("/:opponentId", async (req, res) => {
   await generateNotification(
     finalResult.user._id,
     feedback.message,
-    "Logs",
-    true
+    'Logs',
+    true,
   );
 
   finalResult.user = null;